se agrega en listado de usuario al home, se modifica el toate por sweetalert2

+++ client/src/component/UserCard.js
@@ -1,6 +1,7 @@
 import React from "react";
 
-import toast from "react-hot-toast";
+import Swal from "sweetalert2";
+
 import { useUsers } from "../context/usersContext";
 import { useNavigate } from "react-router-dom";
 
@@ -8,42 +9,25 @@ export default function UserCard({ user }) {
   const navigate = useNavigate();
   const { deleteUser,logout,userRegister } = useUsers();
   const handelDelete = (id, email) => {
-    toast(
-      (t) => (
-        <div>
-          <p className="text-white">
-            Realmente quieres eliminar al usuario - <strong>{email}</strong>
-          </p>
-          <div className="flex justify-around py-2">
-            <button
-              className="bg-red-600 hover:bg-red-400 px-3 py-2 mx-2 rounded-sm text-white text-sm"
-              onClick={() => {
-                deleteUser(id);
-                //verifico si el usuario que elimino es el que esta logueado, si es correcto, elimono al usuario y lo deslogueo
+
+    Swal.fire({
+      title: `'Esta seguro de querear eliminar el usuario ${email}?'`,
+      showDenyButton: true,
+      showCancelButton: true,
+      confirmButtonText: 'Eliminar',
+      // denyButtonText: `canc`,
+    }).then((result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isConfirmed) {
+        deleteUser(id);
+       // verifico si el usuario que elimino es el que esta logueado, si es correcto, elimino al usuario y lo deslogueo
                 if(userRegister.id === id){
                   logout()
-                }
-                toast.dismiss(t.id);
-              }}
-            >
-              Delete
-            </button>
-            <button
-              className="bg-slate-400 hover:bg-slate-500 px-3 py-2 text-white rounded-sm mx-2 text-sm"
-              onClick={() => toast.dismiss(t.id)}
-            >
-              Cancel
-            </button>
-          </div>
-        </div>
-      ),
-      {
-        style: {
-          background: "#202020",
-        },
-        icon: "🎃",
+        Swal.fire('Eliminado!', '', 'success')
       }
-    );
+    }
+    })
+ 
   };
 
   return (
@@ -51,10 +35,10 @@ export default function UserCard({ user }) {
       <div className="m-auto  max-w-lg items-center justify-center overflow-hidden rounded-2xl bg-slate-200 shadow-xl">
         <div className="h-24 bg-white"></div>
         <div className="-mt-20 flex justify-center">
-          <img className="h-32 rounded-full" src={user.image} />
+          <img className="bject-cover w-32 h-32 rounded-full ring-4 ring-teal-500" src={user.image} />
         </div>
         <div className="mt-5 mb-1 px-3 text-center text-lg">{user.name}</div>
-        <div className="mb-5 px-3 text-center text-sky-500">{user.email}</div>
+        <div className="mb-5 px-3 text-center text-teal-500">{user.email}</div>
         <blockquote>
           <p className="mx-2 mb-5 text-center text-base">id - {user.id}</p>
         </blockquote>
